requirements.txt
# Django e extensões
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-rq>=2.8.0
django-allauth>=0.57.0
django-cors-headers>=4.2.0
pyjwt[crypto]

# Documentação da API
drf-spectacular>=0.29.0
drf-spectacular-sidecar>=2024.4.0
drf-spectacular-websocket>=1.3.1

# Channels e WebSockets
channels>=4.0.0
channels-redis>=4.1.0

# Scraping
Scrapy>=2.9.0
scrapy-splash>=0.9.0

# HTTP e networking
requests>=2.31.0

# Redis
redis>=4.5.0

# Async ASGI
asgiref>=3.7.0
websockets>=11.0.0
daphne>=4.2.1

# Variáveis de ambiente
python-decouple>=3.8

